Creation of Auxetic Meshes in MATLAB for GCode Output.
Two methods are represented here; (a) working with sketch projection in SolidWorks and exporting 3D sketch points to Excel, which are then filtered, formatted, and output to GCode in the MATLAB, and (b) doing the whole process in the MATLAB. The MATLAB script can achieve the exact same GCode as the combination of SolidWorks Macros, Excel and the MATLAB, but allows for higher resolution of the traces and faster GCode generation, along with an obviously streamlined process. The reason for using SolidWorks at all was because I was lazy with the math and couldn't see the forest through the trees (3D point capture in Solidworks is already a known and well-documented process).